Removed query cache from readonly monitor live query parsing added executor to refresh index
