+++ lucene/monitor/src/java/org/apache/lucene/monitor/ReadonlyQueryIndex.java
@@ -18,16 +18,19 @@
 package org.apache.lucene.monitor;
 
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import org.apache.lucene.search.IndexSearcher;
-import org.apache.lucene.search.Query;
-import org.apache.lucene.search.SearcherManager;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import org.apache.lucene.index.LeafReaderContext;
+import org.apache.lucene.search.*;
 import org.apache.lucene.store.Directory;
+import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.IOUtils;
+import org.apache.lucene.util.NamedThreadFactory;
 
 class ReadonlyQueryIndex extends QueryIndex {
+
+  private final ScheduledExecutorService refreshExecutor;
 
   public ReadonlyQueryIndex(MonitorConfiguration configuration) throws IOException {
     if (configuration.getDirectoryProvider() == null) {
@@ -35,11 +38,23 @@ class ReadonlyQueryIndex extends QueryIndex {
           "You must specify a Directory when configuring a Monitor as read-only.");
     }
     Directory directory = configuration.getDirectoryProvider().get();
-    this.queries = new HashMap<>();
     this.manager = new SearcherManager(directory, new TermsHashBuilder(termFilters));
     this.decomposer = configuration.getQueryDecomposer();
     this.serializer = configuration.getQuerySerializer();
-    this.populateQueryCache(serializer, decomposer);
+    this.refreshExecutor =
+        Executors.newSingleThreadScheduledExecutor(new NamedThreadFactory("cache-purge"));
+    long refreshFrequency = configuration.getPurgeFrequency();
+    this.refreshExecutor.scheduleAtFixedRate(
+        () -> {
+          try {
+            manager.maybeRefresh();
+          } catch (IOException e) {
+            throw new RuntimeException(e);
+          }
+        },
+        refreshFrequency,
+        refreshFrequency,
+        configuration.getPurgeFrequencyUnits());
   }
 
   @Override
@@ -48,17 +63,19 @@ class ReadonlyQueryIndex extends QueryIndex {
   }
 
   @Override
-  long search(final Query query, QueryCollector matcher) throws IOException {
-    QueryBuilder builder = termFilter -> query;
-    return search(builder, matcher);
-  }
-
-  @Override
   public long search(QueryBuilder queryBuilder, QueryCollector matcher) throws IOException {
     IndexSearcher searcher = null;
     try {
       searcher = manager.acquire();
-      return searchInMemory(queryBuilder, matcher, searcher, this.queries);
+      LazyMonitorQueryCollector collector =
+          new LazyMonitorQueryCollector(matcher, serializer, decomposer);
+      long buildTime = System.nanoTime();
+      Query query =
+          queryBuilder.buildQuery(
+              termFilters.get(searcher.getIndexReader().getReaderCacheHelper().getKey()));
+      buildTime = System.nanoTime() - buildTime;
+      searcher.search(query, collector);
+      return buildTime;
     } finally {
       if (searcher != null) {
         manager.release(searcher);
@@ -67,21 +84,18 @@ class ReadonlyQueryIndex extends QueryIndex {
   }
 
   @Override
-  public void purgeCache() throws IOException {
-    this.populateQueryCache(serializer, decomposer);
-    lastPurged = System.nanoTime();
+  public void purgeCache() {
+    throw new IllegalStateException("Monitor is readOnly, it has no cache");
   }
 
   @Override
-  void purgeCache(CachePopulator populator) throws IOException {
-    manager.maybeRefresh();
-    final Map<String, QueryCacheEntry> newCache = new HashMap<>();
-    populator.populateCacheWithIndex(newCache);
-    queries = newCache;
+  void purgeCache(CachePopulator populator) {
+    throw new IllegalStateException("Monitor is readOnly, it has no cache");
   }
 
   @Override
   public void close() throws IOException {
+    refreshExecutor.shutdown();
     IOUtils.close(manager);
   }
 
@@ -101,6 +115,11 @@ class ReadonlyQueryIndex extends QueryIndex {
   }
 
   @Override
+  public int cacheSize() {
+    return -1;
+  }
+
+  @Override
   public void deleteQueries(List<String> ids) throws IOException {
     throw new IllegalStateException("Monitor is readOnly cannot delete queries");
   }
@@ -111,7 +130,65 @@ class ReadonlyQueryIndex extends QueryIndex {
   }
 
   @Override
+  public long getLastPurged() {
+    return -1;
+  }
+
+  @Override
   public void addListener(MonitorUpdateListener listener) {
     throw new IllegalStateException("Monitor is readOnly cannot register listeners");
+  }
+
+  // ---------------------------------------------
+  //  Helper classes...
+  // ---------------------------------------------
+
+  /** A Collector that decodes the stored query for each document hit reparsing them everytime. */
+  static final class LazyMonitorQueryCollector extends SimpleCollector {
+    private final QueryIndex.QueryCollector matcher;
+    private final QueryIndex.DataValues dataValues = new QueryIndex.DataValues();
+    private final MonitorQuerySerializer serializer;
+    private final QueryDecomposer decomposer;
+
+    LazyMonitorQueryCollector(
+        QueryIndex.QueryCollector matcher,
+        MonitorQuerySerializer serializer,
+        QueryDecomposer decomposer) {
+      this.matcher = matcher;
+      this.serializer = serializer;
+      this.decomposer = decomposer;
+    }
+
+    @Override
+    public void setScorer(Scorable scorer) {
+      this.dataValues.scorer = scorer;
+    }
+
+    @Override
+    public void collect(int doc) throws IOException {
+      dataValues.advanceTo(doc);
+      BytesRef cache_id = dataValues.cacheId.lookupOrd(dataValues.cacheId.ordValue());
+      BytesRef query_id = dataValues.queryId.lookupOrd(dataValues.queryId.ordValue());
+      MonitorQuery mq = serializer.deserialize(dataValues.mq.binaryValue());
+      QueryCacheEntry query =
+          QueryCacheEntry.decompose(mq, decomposer).stream()
+              .filter(queryCacheEntry -> queryCacheEntry.cacheId.equals(cache_id.utf8ToString()))
+              .findFirst()
+              .orElse(null);
+      matcher.matchQuery(query_id.utf8ToString(), query, dataValues);
+    }
+
+    @Override
+    public void doSetNextReader(LeafReaderContext context) throws IOException {
+      this.dataValues.cacheId = context.reader().getSortedDocValues(QueryIndex.FIELDS.cache_id);
+      this.dataValues.queryId = context.reader().getSortedDocValues(QueryIndex.FIELDS.query_id);
+      this.dataValues.mq = context.reader().getBinaryDocValues(QueryIndex.FIELDS.mq);
+      this.dataValues.ctx = context;
+    }
+
+    @Override
+    public ScoreMode scoreMode() {
+      return matcher.scoreMode();
+    }
   }
 }
